Исправил ошибки в модуле  6.1
