+++ js/main.js
@@ -3,5 +3,5 @@ import './util.js';
 import './data.js';
 import './generate-form.js';
 import './check-option-form.js';
-//import './create-map.js';
-import './page-off.js';
+import './activation-map.js';
+import './page.js';
